+++ src/PokeAByte.Domain/Models/Mappers/MapperFileModel.cs
@@ -2,6 +2,12 @@
 
 public record MapperFileModel
 {
+    public MapperFileModel(string id, string displayName)
+    {
+        Id = id;
+        DisplayName = displayName;
+    }
+
     public string Id { get; init; } = string.Empty;
     public string DisplayName { get; init; } = string.Empty;